# QuantumCircuit cache bits