--- a/qiskit/circuit/library/blueprintcircuit.py
+++ b/qiskit/circuit/library/blueprintcircuit.py
@@ -41,6 +41,7 @@ class BlueprintCircuit(QuantumCircuit, ABC):
         self._data = None
         self._qregs = []
         self._cregs = []
+        self._qubits = []
 
     @abstractmethod
     def _check_configuration(self, raise_on_failure: bool = True) -> bool:
@@ -82,6 +83,7 @@ class BlueprintCircuit(QuantumCircuit, ABC):
     def qregs(self, qregs):
         """Set the quantum registers associated with the circuit."""
         self._qregs = qregs
+        self._qubits = [qbit for qreg in qregs for qbit in qreg]
         self._invalidate()
 
     @property